divide project into 3 units (project,app 1, app 2)
