+++ templates/base.html
@@ -42,10 +42,10 @@
               <div class="dropdown-menu" aria-labelledby="dropdown09">
                 <a class="dropdown-item" href="#">About</a>
                 {% if request.user.is_authenticated %}
-                <a class="dropdown-item" href="{% url 'notes:logout' %}">Logout</a>
+                <a class="dropdown-item" href="{% url 'logout' %}">Logout</a>
                 {% else %}
-                  <a class="dropdown-item" href="{% url 'notes:register' %}">register</a>
-                  <a class="dropdown-item" href="{% url 'notes:login' %}">login</a>
+                  <a class="dropdown-item" href="{% url 'register' %}">register</a>
+                  <a class="dropdown-item" href="{% url 'login' %}">login</a>
                 {% endif %}
               </div>
             </li>
