+++ templates/login.html
@@ -48,7 +48,7 @@
                             <button class="btn btn-lg btn-primary btn-block" type="submit">Sign in</button>
                             <p class="mt-5 mb-3 text-muted text-center">© 2017-2018</p>
                 </form>
-                <p>dont have an account?  <a href="{% url 'notes:register' %}">register</a> </p>
+                <p>dont have an account?  <a href="{% url 'register' %}">register</a> </p>
             </div>
     </div>
 </body>
