+++ templates/signup.html
@@ -38,7 +38,7 @@
                             {% endfor %}
                             <input type="submit" class="btn btn-primary" value="submit"/>
                         </form>
-                        <p>already have an account?  <a href="{% url 'notes:login' %}">login</a> </p>
+                        <p>already have an account?  <a href="{% url 'login' %}">login</a> </p>
             </div>
         </div>
     </div>
