+++ todo/urls.py
@@ -1,16 +1,15 @@
 from django.urls import path,include
-
-from .views import home, login_view, register_view, logout_view, note_detail, edit_note, delete_note, LandingView
+from todo import views
 
 app_name = 'notes'
 
 urlpatterns = [
 
     #home page and dashboard
-    path('home', LandingView.as_view() , name='landing' ),
-    path('mine/', home, name='home' ),
-    path('<slug:slug>/<int:pk>/', note_detail, name='note_detail'),
-    path('<slug:slug>/<int:pk>/edit', edit_note, name='edit_note'),
-    path('<slug:slug>/<int:pk>/delete', delete_note, name='note_confirm_delete'),
-
+    path('', views.home, name='home' ),
+    path('<slug:slug>/<int:pk>/', include([
+        path('',views.note_detail, name='note_detail'),
+        path('edit', views.edit_note, name='edit_note'),
+        path('delete', views.delete_note, name='note_confirm_delete')
+        ])),
 ]
